app/requirements.txt
Flask-SQLAlchemy
psycopg2-binary